feat: refactor bank transactions section
