+++ app/src/views/SingleTeamView.vue
@@ -72,7 +72,7 @@
           />
         </template>
         <template #tab-2>
-          <BankTransactions v-if="activeTab == 2" :bank-address="team.bankAddress" />
+          <BankTransactionsSection v-if="activeTab == 2" :bank-address="team.bankAddress" />
         </template>
         <template #tab-3>
           <ProposalDashBoard :team="team" @getTeam="getTeamAPI" />
@@ -123,7 +123,7 @@ import AddMemberCard from '@/components/AddMemberCard.vue'
 import TeamDetails from '@/components/sections/SingleTeamView/Team/TeamDetails.vue'
 import ModalComponent from '@/components/ModalComponent.vue'
 import TabNavigation from '@/components/TabNavigation.vue'
-import BankTransactions from '@/components/sections/SingleTeamView/BankTransaction/BankTransactions.vue'
+import BankTransactions from '@/components/sections/SingleTeamView/BankTransactionsSection.vue'
 import TeamAccount from '@/components/sections/SingleTeamView/Bank/TeamAccount.vue'
 import ProposalDashBoard from '@/components/sections/SingleTeamView/Governance/ProposalDashboard.vue'
